Inline AuthenticateDialog inside LoginDialog
There are no other subclasses of it anyway
