+++ app/src/main/kotlin/com/hedvig/app/authenticate/LoginDialog.kt
@@ -1,19 +1,46 @@
 package com.hedvig.app.authenticate
 
+import android.app.Dialog
+import android.content.Intent
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
+import android.net.Uri
 import android.os.Bundle
+import android.view.LayoutInflater
 import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.DialogFragment
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
+import com.hedvig.app.databinding.DialogAuthenticateBinding
 import com.hedvig.app.feature.genericauth.GenericAuthActivity
 import com.hedvig.app.feature.loggedin.ui.LoggedInActivity
+import com.hedvig.app.util.QR
+import com.hedvig.app.util.extensions.canOpenUri
 import com.hedvig.authlib.LoginStatusResult
+import com.zhuinden.fragmentviewbindingdelegatekt.viewBinding
 import hedvig.resources.R
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class LoginDialog : AuthenticateDialog() {
+class LoginDialog : DialogFragment() {
+
+  val binding by viewBinding(DialogAuthenticateBinding::bind)
   private val viewModel: UserViewModel by viewModel()
+
+  override fun onCreateView(
+    inflater: LayoutInflater,
+    container: ViewGroup?,
+    savedInstanceState: Bundle?,
+  ): View? = inflater.inflate(com.hedvig.app.R.layout.dialog_authenticate, container, false)
+
+  override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
+    return super.onCreateDialog(savedInstanceState).apply {
+      window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
+      setCanceledOnTouchOutside(false)
+    }
+  }
 
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
     super.onViewCreated(view, savedInstanceState)
@@ -29,6 +56,33 @@ class LoginDialog : AuthenticateDialog() {
 
     binding.login.setOnClickListener {
       requireActivity().startActivity(GenericAuthActivity.newInstance(requireActivity()))
+    }
+  }
+
+  private fun handleAutoStartToken(autoStartToken: String) {
+    val autoStartUrl = "bankid:///?autostarttoken=$autoStartToken"
+    val bankIdUri = Uri.parse("$autoStartUrl&redirect=null")
+    if (requireContext().canOpenUri(bankIdUri)) {
+      startActivity(
+        Intent(
+          Intent.ACTION_VIEW,
+          bankIdUri,
+        ),
+      )
+    } else {
+      QR
+        .with(requireContext())
+        .load(autoStartUrl)
+        .into(binding.qrCode)
+    }
+  }
+
+  fun redirect() {
+    val bankIdUri = Uri.parse("bankid://?redirectUrl=hedvig://")
+    if (requireContext().canOpenUri(bankIdUri)) {
+      val intent = Intent(Intent.ACTION_VIEW, bankIdUri)
+      intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
+      startActivity(intent)
     }
   }
